# Update inmem pipe to use go channel