The inmemory pipe implementation may simply use go channels instead of using a pipe.

Closes #92
Alternate implementation to #100