@dmcgowan 2015-09-16
@choleraehyq have a look at this as well. Sorry I really should have put this in a PR sooner instead of let it sit for months. This does add a buffered pipe as well which is kind of a new concept in libchan but I think it could be powerful since they are just as capable of being sent over the wire.