@choleraehyq 2015-09-18
@dmcgowan 
#100 has the same API as before, so it may have a better compatibility. Except for that, I think your version is more elegant.