@dmcgowan 2015-10-02
@choleraehyq this does not change any of the exposed API except to add a BufferedPipe.